Move existing layer API to separate module, move shared code
