+++ libcnb/src/layer/layer_api/handling.rs
@@ -1,11 +1,14 @@
 // This lint triggers when both layer_dir and layers_dir are present which are quite common.
 #![allow(clippy::similar_names)]
 
-use super::public_interface::Layer;
+use super::Layer;
 use crate::build::BuildContext;
 use crate::data::layer::LayerName;
 use crate::data::layer_content_metadata::LayerContentMetadata;
 use crate::generic::GenericMetadata;
+use crate::layer::shared::{
+    delete_layer, DeleteLayerError, ReadLayerError, WriteLayerMetadataError,
+};
 use crate::layer::{ExistingLayerStrategy, LayerData, MetadataMigration};
 use crate::layer_env::LayerEnv;
 use crate::sbom::{cnb_sbom_path, Sbom};
@@ -221,21 +224,6 @@ pub enum HandleLayerError {
 }
 
 #[derive(thiserror::Error, Debug)]
-pub enum DeleteLayerError {
-    #[error("I/O error while deleting existing layer: {0}")]
-    IoError(#[from] std::io::Error),
-}
-
-#[derive(thiserror::Error, Debug)]
-pub enum ReadLayerError {
-    #[error("Layer content metadata couldn't be parsed!")]
-    LayerContentMetadataParseError(toml::de::Error),
-
-    #[error("Unexpected I/O error while reading layer: {0}")]
-    IoError(#[from] std::io::Error),
-}
-
-#[derive(thiserror::Error, Debug)]
 #[allow(clippy::enum_variant_names)]
 pub enum WriteLayerError {
     #[error("{0}")]
@@ -249,15 +237,6 @@ pub enum WriteLayerError {
 
     #[error("{0}")]
     ReplaceLayerSbomsError(#[from] ReplaceLayerSbomsError),
-}
-
-#[derive(thiserror::Error, Debug)]
-pub enum WriteLayerMetadataError {
-    #[error("Unexpected I/O error while writing layer metadata: {0}")]
-    IoError(#[from] std::io::Error),
-
-    #[error("Error while writing layer content metadata TOML: {0}")]
-    TomlFileError(#[from] TomlFileError),
 }
 
 #[derive(thiserror::Error, Debug)]
@@ -291,20 +270,6 @@ pub(in crate::layer) enum ExecDPrograms {
 pub(in crate::layer) enum Sboms {
     Keep,
     Replace(Vec<Sbom>),
-}
-
-/// Does not error if the layer doesn't exist.
-pub(in crate::layer) fn delete_layer<P: AsRef<Path>>(
-    layers_dir: P,
-    layer_name: &LayerName,
-) -> Result<(), DeleteLayerError> {
-    let layer_dir = layers_dir.as_ref().join(layer_name.as_str());
-    let layer_toml = layers_dir.as_ref().join(format!("{layer_name}.toml"));
-
-    default_on_not_found(remove_dir_recursively(&layer_dir))?;
-    default_on_not_found(fs::remove_file(layer_toml))?;
-
-    Ok(())
 }
 
 pub(in crate::layer) fn replace_layer_sboms<P: AsRef<Path>>(
@@ -420,81 +385,20 @@ pub(crate) fn read_layer<M: DeserializeOwned, P: AsRef<Path>>(
     layers_dir: P,
     layer_name: &LayerName,
 ) -> Result<Option<LayerData<M>>, ReadLayerError> {
-    let layer_dir_path = layers_dir.as_ref().join(layer_name.as_str());
-    let layer_toml_path = layers_dir.as_ref().join(format!("{layer_name}.toml"));
+    crate::layer::shared::read_layer(layers_dir, layer_name).map(|read_layer| {
+        read_layer.map(|read_layer| {
 
-    if !layer_dir_path.exists() && !layer_toml_path.exists() {
-        return Ok(None);
-    } else if !layer_dir_path.exists() && layer_toml_path.exists() {
-        // This is a valid case according to the spec:
-        // https://github.com/buildpacks/spec/blob/7b20dfa070ed428c013e61a3cefea29030af1732/buildpack.md#layer-types
-        //
-        // When launch = true, build = false, cache = false, the layer metadata will be restored but
-        // not the layer itself. However, we choose to not support this case as of now. It would
-        // complicate the API we need to expose to the user of libcnb as this case is very different
-        // compared to all other combinations of launch, build and cache. It's the only case where
-        // a cache = false layer restores some of its data between builds.
-        //
-        // To normalize, we remove the layer TOML file and treat the layer as non-existent.
-        fs::remove_file(&layer_toml_path)?;
-        return Ok(None);
-    }
-
-    // An empty layer content metadata file is valid and the CNB spec is not clear if the lifecycle
-    // has to restore them if they're empty. This is especially important since the layer types
-    // are removed from the file if it's restored. To normalize, we write an empty file if the layer
-    // directory exists without the metadata file.
-    if !layer_toml_path.exists() {
-        fs::write(&layer_toml_path, "")?;
-    }
-
-    let layer_toml_contents = fs::read_to_string(&layer_toml_path)?;
-
-    let layer_content_metadata = toml::from_str::<LayerContentMetadata<M>>(&layer_toml_contents)
-        .map_err(ReadLayerError::LayerContentMetadataParseError)?;
-
-    let layer_env = LayerEnv::read_from_layer_dir(&layer_dir_path)?;
-
-    Ok(Some(LayerData {
-        name: layer_name.clone(),
-        path: layer_dir_path,
-        env: layer_env,
-        content_metadata: layer_content_metadata,
-    }))
-}
-
-pub(crate) fn replace_layer_types<P: AsRef<Path>>(
-    layers_dir: P,
-    layer_name: &LayerName,
-    layer_types: LayerTypes,
-) -> Result<(), WriteLayerMetadataError> {
-    let layer_content_metadata_path = layers_dir.as_ref().join(format!("{layer_name}.toml"));
-
-    let mut content_metadata =
-        read_toml_file::<LayerContentMetadata>(&layer_content_metadata_path)?;
-    content_metadata.types = Some(layer_types);
-
-    write_toml_file(&content_metadata, &layer_content_metadata_path)
-        .map_err(WriteLayerMetadataError::TomlFileError)
-}
-
-pub(crate) fn replace_layer_metadata<M: Serialize, P: AsRef<Path>>(
-    layers_dir: P,
-    layer_name: &LayerName,
-    metadata: M,
-) -> Result<(), WriteLayerMetadataError> {
-    let layer_content_metadata_path = layers_dir.as_ref().join(format!("{layer_name}.toml"));
-
-    let content_metadata = read_toml_file::<LayerContentMetadata>(&layer_content_metadata_path)?;
-
-    write_toml_file(
-        &LayerContentMetadata {
-            types: content_metadata.types,
-            metadata,
-        },
-        &layer_content_metadata_path,
-    )
-    .map_err(WriteLayerMetadataError::TomlFileError)
+            LayerEnv::read_from_layer_dir(&read_layer.path).map_err(ReadLayerError::IoError)
+            
+            
+            LayerEnv::read_from_layer_dir(&read_layer.path).map(|layer_env| LayerData {
+                name: read_layer.name,
+                path: read_layer.path,
+                content_metadata: read_layer.metadata,
+                env: layer_env,
+            })
+        })
+    })
 }
 
 #[cfg(test)]
