+++ libcnb/src/layer/struct_api/execute.rs
@@ -1,9 +1,10 @@
-use crate::layer::handling::{ExecDPrograms, Sboms};
-use crate::layer::{
-    DeleteLayerError, EmptyReason, InspectExistingAction, IntoAction, InvalidMetadataAction,
-    LayerContents, LayerRef, ReadLayerError, WriteLayerError,
+use crate::layer::shared::{
+    delete_layer, read_layer, replace_layer_metadata, replace_layer_types, DeleteLayerError,
+    ReadLayerError, WriteLayerError,
 };
-use crate::layer_env::LayerEnv;
+use crate::layer::{
+    EmptyReason, InspectExistingAction, IntoAction, InvalidMetadataAction, LayerContents, LayerRef,
+};
 use crate::Buildpack;
 use libcnb_common::toml_file::{read_toml_file, TomlFileError};
 use libcnb_data::generic::GenericMetadata;
@@ -43,17 +44,16 @@ where
     MA: IntoAction<InvalidMetadataAction<M>, MC, B::Error>,
     IA: IntoAction<InspectExistingAction, IC, B::Error>,
 {
-    match crate::layer::handling::read_layer::<M, _>(layers_dir, &layer_name) {
+    match read_layer::<M, _>(layers_dir, &layer_name) {
         Ok(None) => create_layer(layer_types, &layer_name, layers_dir, EmptyReason::Uncached),
         Ok(Some(layer_data)) => {
-            let inspect_action =
-                inspect_existing(&layer_data.content_metadata.metadata, &layer_data.path)
-                    .into_action()
-                    .map_err(crate::Error::BuildpackError)?;
+            let inspect_action = inspect_existing(&layer_data.metadata.metadata, &layer_data.path)
+                .into_action()
+                .map_err(crate::Error::BuildpackError)?;
 
             match inspect_action {
                 (InspectExistingAction::Delete, cause) => {
-                    crate::layer::handling::delete_layer(layers_dir, &layer_name)
+                    delete_layer(layers_dir, &layer_name)
                         .map_err(ExecuteLayerDefinitionError::DeleteLayerError)?;
 
                     create_layer(
@@ -67,12 +67,11 @@ where
                     // Always write the layer types as:
                     // a) they might be different from what is currently on disk
                     // b) the cache field will be removed by CNB lifecycle on cache restore
-                    crate::layer::replace_layer_types(layers_dir, &layer_name, layer_types)
-                        .map_err(|error| {
-                            ExecuteLayerDefinitionError::WriteLayerError(
-                                WriteLayerError::WriteLayerMetadataError(error),
-                            )
-                        })?;
+                    replace_layer_types(layers_dir, &layer_name, layer_types).map_err(|error| {
+                        ExecuteLayerDefinitionError::WriteLayerError(
+                            WriteLayerError::WriteLayerMetadataError(error),
+                        )
+                    })?;
 
                     Ok(LayerRef {
                         name: layer_data.name,
@@ -95,7 +94,7 @@ where
 
             match invalid_metadata_action {
                 (InvalidMetadataAction::DeleteLayer, cause) => {
-                    crate::layer::handling::delete_layer(layers_dir, &layer_name)
+                    delete_layer(layers_dir, &layer_name)
                         .map_err(ExecuteLayerDefinitionError::DeleteLayerError)?;
 
                     create_layer(
@@ -106,12 +105,11 @@ where
                     )
                 }
                 (InvalidMetadataAction::ReplaceMetadata(metadata), _) => {
-                    crate::layer::replace_layer_metadata(layers_dir, &layer_name, metadata)
-                        .map_err(|error| {
-                            ExecuteLayerDefinitionError::WriteLayerError(
-                                WriteLayerError::WriteLayerMetadataError(error),
-                            )
-                        })?;
+                    replace_layer_metadata(layers_dir, &layer_name, metadata).map_err(|error| {
+                        ExecuteLayerDefinitionError::WriteLayerError(
+                            WriteLayerError::WriteLayerMetadataError(error),
+                        )
+                    })?;
 
                     execute(
                         layer_types,
@@ -138,25 +136,21 @@ fn create_layer<X, Y, B>(
 where
     B: Buildpack + ?Sized,
 {
-    crate::layer::handling::write_layer(
+    crate::layer::shared::write_layer(
         layers_dir,
         layer_name,
-        &LayerEnv::new(),
         &LayerContentMetadata {
             types: Some(layer_types),
             metadata: GenericMetadata::default(),
         },
-        ExecDPrograms::Keep,
-        Sboms::Keep,
     )
     .map_err(ExecuteLayerDefinitionError::WriteLayerError)?;
 
-    let layer_data =
-        crate::layer::handling::read_layer::<GenericMetadata, _>(layers_dir, layer_name)
-            .map_err(ExecuteLayerDefinitionError::ReadLayerError)?
-            .ok_or(ExecuteLayerDefinitionError::CouldNotReadLayerAfterCreate(
-                layer_name.clone(),
-            ))?;
+    let layer_data = read_layer::<GenericMetadata, _>(layers_dir, layer_name)
+        .map_err(ExecuteLayerDefinitionError::ReadLayerError)?
+        .ok_or(ExecuteLayerDefinitionError::CouldNotReadLayerAfterCreate(
+            layer_name.clone(),
+        ))?;
 
     Ok(LayerRef {
         name: layer_data.name,
